feat: Icono frutas

+++ app/src/main/java/com/example/frutapp/data/model/NutritionsDTO.kt
@@ -1,9 +1,9 @@
 package com.example.frutapp.data.model
 
 data class NutritionsDTO(
-  val carbohydrates: Float,
-  val protein: Float,
-  val fat: Float,
+  val carbohydrates: Double,
+  val protein: Double,
+  val fat: Double,
   val calories: Int,
-  val sugar: Float
+  val sugar: Double
 )
